Blackjack domain from AAAI 2014 paper.

+++ src/edu/oregonstate/eecs/mcplan/domains/voyager/VoyagerInstance.java
@@ -192,9 +192,9 @@ public class VoyagerInstance extends Instance<VoyagerInstance, VoyagerState, Voy
 	}
 	
 	@Override
-	public long nextSeed()
+	public int nextSeed()
 	{
-		return rng_.nextLong();
+		return rng_.nextInt();
 	}
 
 	@Override
